+++ src/ISettlerActions.sol
@@ -255,6 +255,17 @@ interface ISettlerActions {
         uint256 amountOutMin
     ) external;
 
+    function EKUBOV3(
+        address recipient,
+        address sellToken,
+        uint256 bps,
+        bool feeOnTransfer,
+        uint256 hashMul,
+        uint256 hashMod,
+        bytes memory fills,
+        uint256 amountOutMin
+    ) external;
+
     function EKUBO_VIP(
         address recipient,
         ISignatureTransfer.PermitTransferFrom memory permit,
@@ -266,7 +277,28 @@ interface ISettlerActions {
         uint256 amountOutMin
     ) external;
 
+    function EKUBOV3_VIP(
+        address recipient,
+        ISignatureTransfer.PermitTransferFrom memory permit,
+        bool feeOnTransfer,
+        uint256 hashMul,
+        uint256 hashMod,
+        bytes memory fills,
+        bytes memory sig,
+        uint256 amountOutMin
+    ) external;
+
     function METATXN_EKUBO_VIP(
+        address recipient,
+        ISignatureTransfer.PermitTransferFrom memory permit,
+        bool feeOnTransfer,
+        uint256 hashMul,
+        uint256 hashMod,
+        bytes memory fills,
+        uint256 amountOutMin
+    ) external;
+
+    function METATXN_EKUBOV3_VIP(
         address recipient,
         ISignatureTransfer.PermitTransferFrom memory permit,
         bool feeOnTransfer,